Refactor annotations: Relocated `Password`, `PasswordMatches`, and `ValidIban` to `shared.annotations` package for improved reusability.
