+++ src/main/java/info/mackiewicz/bankapp/presentation/dashboard/dto/InternalTransferRequest.java
@@ -1,7 +1,7 @@
 package info.mackiewicz.bankapp.presentation.dashboard.dto;
 
 import info.mackiewicz.bankapp.account.validation.DifferentAccounts;
-import info.mackiewicz.bankapp.account.validation.ValidIban;
+import info.mackiewicz.bankapp.shared.annotations.ValidIban;
 import info.mackiewicz.bankapp.transaction.model.TransactionType;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotEmpty;
