+++ src/main/java/info/mackiewicz/bankapp/shared/annotations/PasswordMatches.java
@@ -1,10 +1,12 @@
-package info.mackiewicz.bankapp.presentation.auth.validation;
+package info.mackiewicz.bankapp.shared.annotations;
 
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
+
 import java.lang.annotation.Documented;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
+
 import static java.lang.annotation.ElementType.TYPE;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
