+++ src/main/java/info/mackiewicz/bankapp/shared/annotations/PasswordMatchesValidator.java
@@ -1,7 +1,6 @@
-package info.mackiewicz.bankapp.presentation.auth.validation;
+package info.mackiewicz.bankapp.shared.annotations;
 
 import info.mackiewicz.bankapp.shared.web.dto.interfaces.PasswordConfirmation;
-
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
