+++ src/main/java/info/mackiewicz/bankapp/system/banking/operations/controller/dto/TransactionRequest.java
@@ -2,7 +2,7 @@ package info.mackiewicz.bankapp.system.banking.operations.controller.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import info.mackiewicz.bankapp.account.validation.ValidIban;
+import info.mackiewicz.bankapp.shared.annotations.ValidIban;
 import info.mackiewicz.bankapp.transaction.model.Transaction;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.media.Schema.RequiredMode;
